update header to show username
