--- web/src/views/HomeView.vue
@@ -1,15 +1,0 @@
-<template>
-  <UserHome/>
-</template>
-
-<script>
-import UserHome from "../components/UserHome.vue"
-export default {
-  components: { UserHome }
-
-}
-</script>
-
-<style>
-
-</style>
